+++ docs/release-notes.md
@@ -6,6 +6,8 @@
 
 - Allow submitting paused tasks.
 - Allow changing in-place attributes of tasks when submitting with `add_task`.
+- Allow selecting logger (classical, loguru).
+- Allow naming schedulers with an extra logger name.
 
 ### Fixed
 
@@ -18,6 +20,7 @@
 - `pending_tasks` has now no more store alias in it.
 - `tzlocal` is now optional.
 - Tasks use the timezone of the scheduler for their triggers which require a timezone.
+- `loguru` is now optional.
 
 ## 0.10.1